refactor(controller): 아직 포장되지 않은 원시값을 dto로 포장하여 전달할 수 있도록 구현
* LottoController.java: 아직 포장되지 않은 원시값들을 dto로 포장하여 전달할 수 있도록 로직 수정
* LottoResponse.java: List<Integer> 값을 dto로 전달하도록 dto 생성
* ReturnRateResponse.java: 수익률을 dto로 포장하여 전달하도록 dto 생성
