+++ src/main/java/lotto/mapper/LottoMapper.java
@@ -1,11 +1,15 @@
-package lotto.domain.mapper;
+package lotto.mapper;
 
 import java.util.Map.Entry;
-import lotto.domain.rank.LottoPrice;
+import lotto.domain.LottoPrice;
+import lotto.domain.Number;
+import lotto.domain.lotto.Lotto;
 import lotto.domain.util.LottoStatistics;
 import lotto.view.dto.LottoCountResponse;
 import lotto.view.dto.LottoErrorResponse;
 import lotto.view.dto.LottoRankResponse;
+import lotto.view.dto.LottoResponse;
+import lotto.view.dto.ReturnRateResponse;
 
 public class LottoMapper {
 
@@ -24,5 +28,18 @@ public class LottoMapper {
 
     public static LottoCountResponse toLottoCountResponse(int lottoCount) {
         return new LottoCountResponse(lottoCount);
+    }
+
+    public static ReturnRateResponse toReturnRateResponse(double returnRate) {
+        return new ReturnRateResponse(returnRate);
+    }
+
+    public static LottoResponse toLottoResponse(Lotto lotto) {
+        return new LottoResponse(
+                lotto.displayLotto()
+                        .stream()
+                        .map(Number::getNumber)
+                        .toList()
+        );
     }
 }
