+++ src/main/java/lotto/view/dto/LottoResponse.java
@@ -1,0 +1,8 @@
+package lotto.view.dto;
+
+import java.util.List;
+
+public record LottoResponse(
+        List<Integer> lottoNumber
+) {
+}
