+++ src/main/java/lotto/view/dto/ReturnRateResponse.java
@@ -1,0 +1,6 @@
+package lotto.view.dto;
+
+public record ReturnRateResponse(
+        double returnRate
+) {
+}
